requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
firebase-admin==6.5.0
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
httpx==0.27.0

# ── Authentication ──
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.9

# ── Frontend (Jinja2 templates + static files) ──
jinja2==3.1.4
aiofiles==23.2.1
